requirements.txt
python-telegram-bot==21.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
APScheduler==3.10.4
python-telegram-bot-calendar==1.0.5
